Alter the mapping list element struct to use 4 bytes to store the protocol instead of 1. Despite only needing at most 2 bits this prevents bus errors on strict alignment platforms such as armish and sparc64.

+++ natpmpd.c
@@ -75,7 +75,7 @@ struct timeval timeouts[NATPMPD_MAX_DELAY] = {
 };
 
 struct mapping {
-	u_int8_t		 proto;
+	u_int32_t		 proto;
 	struct sockaddr		 dst;
 	struct sockaddr		 rdr;
 	struct event		 ev;
